test/topology: split tablets tests
Split tablets test cases to own test file as the total running time is
too long (14 minutes).

Refs #13905
