+++ test/topology_experimental_raft/test_tablets_scans.py
@@ -1,0 +1,36 @@
+#
+# Copyright (C) 2023-present ScyllaDB
+#
+# SPDX-License-Identifier: AGPL-3.0-or-later
+#
+
+import asyncio
+import logging
+from test.pylib.manager_client import ManagerClient
+import pytest
+
+logger = logging.getLogger(__name__)
+
+
+@pytest.mark.asyncio
+async def test_scans(manager: ManagerClient):
+    logger.info("Bootstrapping cluster")
+    servers = [await manager.server_add(), await manager.server_add(), await manager.server_add()]
+
+    cql = manager.get_cql()
+    await cql.run_async("CREATE KEYSPACE test WITH replication = {'class': 'NetworkTopologyStrategy', "
+                  "'replication_factor': 1, 'initial_tablets': 8};")
+    await cql.run_async("CREATE TABLE test.test (pk int PRIMARY KEY, c int);")
+
+    keys = range(100)
+    await asyncio.gather(*[cql.run_async(f"INSERT INTO test.test (pk, c) VALUES ({k}, {k});") for k in keys])
+
+    rows = await cql.run_async("SELECT count(*) FROM test.test;")
+    assert rows[0].count == len(keys)
+
+    rows = await cql.run_async("SELECT * FROM test.test;")
+    assert len(rows) == len(keys)
+    for r in rows:
+        assert r.c == r.pk
+
+    await cql.run_async("DROP KEYSPACE test;")
